backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
google-genai>=1.0.0
python-dotenv>=1.0.1
pydantic>=2.10.0
tree-sitter>=0.22.0
tree-sitter-python>=0.23.0
tree-sitter-r>=0.1.0
